+++ packages/dev-require/CHANGELOG.md
@@ -1,5 +1,11 @@
 # @aevans04/dev-require
 
+## 1.1.0
+
+### Minor Changes
+
+- Changed structure of dev-require config file
+
 ## 1.0.0
 
 ### Major Changes