+++ packages/package-b/CHANGELOG.md
@@ -1,5 +1,13 @@
 # @aevans04/test-node-monorepo-b
 
+## 1.0.4
+
+### Patch Changes
+
+- Changed structure of dev-require config file
+- Updated dependencies
+  - @aevans04/dev-require@1.1.0
+
 ## 1.0.3
 
 ### Patch Changes